Code refactoring
- Added a NewsManager component that downloads and processes all the data
