+++ src/components/Home/index.tsx
@@ -3,4 +3,4 @@ export * from "./MainTitle"
 export * from "./MenuButton"
 export * from "./StickyHeader"
 export * from "./PoliSearchBar"
-export * from "./NewsBottomSheet"
+export * from "./News/NewsManager"
